refactor: rename PracticeTest to Writing and OralPractice to Speaking (#124)
