+++ frontend/src/pages/WritingPage.tsx
